+++ dashboard/src/main/home/app-dashboard/validate-apply/app-settings/EnvSettings.tsx
@@ -5,7 +5,6 @@ import { PopulatedEnvGroup } from "./types";
 import EnvVariables from "./EnvVariables";
 import EnvGroups from "./EnvGroups";
 import { AppRevision } from "lib/revisions/types";
-import { DetectedServices } from "lib/porter-apps/services";
 
 type Props = {
   appName?: string;
@@ -18,7 +17,7 @@ type Props = {
 const EnvSettings: React.FC<Props> = (props) => {
   return (
     <>
-      <EnvVariables />
+      <EnvVariables syncedEnvGroups={[]}/>
       <EnvGroups {...props} attachedEnvGroups={props.attachedEnvGroups} />
     </>
   );
